[#12] Checking that applying our config works
jboss-cli rollback the configuration if any line fails, but the script still returns 0.

Greping the results for a result that is different from { "output": "success"}

+++ Dockerfile
@@ -8,7 +8,8 @@ COPY docker-entrypoint.sh ${JB_HOME}
 COPY theme/akvo/ ${KC_HOME}/themes/akvo/
 COPY keycloak_configuration.txt ${KC_HOME}
 
-RUN ${KC_HOME}/bin/jboss-cli.sh --file=${KC_HOME}/keycloak_configuration.txt && \
+RUN ${KC_HOME}/bin/jboss-cli.sh --file=${KC_HOME}/keycloak_configuration.txt | tee ${KC_HOME}/cli-result.txt && \
+    grep outcome ${KC_HOME}/cli-result.txt | grep -v success && exit 123 || \
     rm -rf ${KC_HOME}/standalone/configuration/standalone_xml_history/current/* && \
     mkdir -p ${KC_HOME}/modules/system/layers/base/com/mysql/jdbc/main; \
     cd ${KC_HOME}/modules/system/layers/base/com/mysql/jdbc/main && \
